+++ msl/network/__init__.py
@@ -26,7 +26,7 @@ from .manager import (
 
 __author__ = 'Measurement Standards Laboratory of New Zealand'
 __copyright__ = '\xa9 2017 - 2021, ' + __author__
-__version__ = '0.6.0.dev0'
+__version__ = '1.0.0.dev0'
 
 _v = re.search(r'(\d+)\.(\d+)\.(\d+)[.-]?(.*)', __version__).groups()